udpate: to get charts from async git repo into build
- new make target
- update dockerfile*
- set new flag in main and bit renaming for variable
