+++ Dockerfile.konflux
@@ -14,18 +14,21 @@ COPY cmd/ cmd/
 COPY api/ api/
 COPY internal/ internal/
 
-# Obtain the batch-gateway Helm chart into /tmp/batch-gateway-chart.
+# Obtain Helm charts into /tmp staging directories.
 COPY Makefile ./
 
-RUN mkdir -p -m 777 /tmp/batch-gateway-chart /tmp/prefetched-charts
+RUN mkdir -p -m 777 /tmp/batch-gateway-chart /tmp/async-processor-chart /tmp/prefetched-charts
 # For all the crazy permisison denies
 COPY --chown=1001:0 Dockerfile prefetched-charts* /tmp/prefetched-charts/
 
 RUN if [ "${BUILD_TYPE}" != "CI" ]; then \
         make fetch-batch-gateway && \
-        cp -r batch-gateway/charts/batch-gateway/* /tmp/batch-gateway-chart/; \
+        cp -r batch-gateway/charts/batch-gateway/* /tmp/batch-gateway-chart/ && \
+        make fetch-llm-d-async && \
+        cp -r llm-d-async/charts/async-processor/* /tmp/async-processor-chart/; \
     else \
         cp -r /tmp/prefetched-charts/batch-gateway/* /tmp/batch-gateway-chart/ 2>/dev/null || true; \
+        cp -r /tmp/prefetched-charts/async-processor/* /tmp/async-processor-chart/ 2>/dev/null || true; \
     fi
 
 RUN mkdir -p bin && \
@@ -37,9 +40,12 @@ FROM registry.access.redhat.com/ubi9/ubi-minimal:9.7@sha256:907b68736aa798b2d382
 WORKDIR /
 COPY --from=builder /opt/app-root/src/bin/manager /manager
 COPY --from=builder /tmp/batch-gateway-chart /charts/batch-gateway/
+COPY --from=builder /tmp/async-processor-chart /charts/async-processor/
 RUN rm -f /charts/batch-gateway/Dockerfile && \
     test -f /charts/batch-gateway/Chart.yaml || \
-    { echo "ERROR: /charts/batch-gateway/Chart.yaml not found. Chart directory contents:"; ls -la /charts/batch-gateway/; exit 1; }
+    { echo "ERROR: /charts/batch-gateway/Chart.yaml not found. Chart directory contents:"; ls -la /charts/batch-gateway/; exit 1; } && \
+    test -f /charts/async-processor/Chart.yaml || \
+    { echo "ERROR: /charts/async-processor/Chart.yaml not found. Chart directory contents:"; ls -la /charts/async-processor/; exit 1; }
 
 USER 1001:1001
 ENTRYPOINT ["/manager"]
